# openapi/openapi.json
{
  "components": {
    "schemas": {
...
      "ExecutionTarget": {
        "description": "Execution target.",
        "enum": [
          "pool",
          "on_demand"
        ],
        "title": "ExecutionTarget",
        "type": "string"
      },
...
      "ExperimentRunCreateRequest": {
        "additionalProperties": false,
        "description": "Experiment run create request.",
        "properties": {
          "agent_version_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Id of the agent version to execute, the latest runnable version when omitted.",
            "title": "Agent Version Id"
          },
          "execution_target": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ExecutionTarget"
              },
              {
                "type": "null"
              }
            ],
            "description": "Execution target, the agent version's default when omitted."
          },
          "score_baselines": {
            "default": false,
            "description": "Whether the runner also scores originals missing scores.",
            "title": "Score Baselines",
            "type": "boolean"
          }
        },
        "title": "ExperimentRunCreateRequest",
        "type": "object"
      },
...
      "ExperimentRunResponse": {
        "description": "Experiment run response.",
        "properties": {
          "agent_version_id": {
            "description": "Id of the agent version.",
            "format": "uuid",
            "title": "Agent Version Id",
            "type": "string"
          },
          "created": {
            "description": "Creation time.",
            "format": "date-time",
            "title": "Created",
            "type": "string"
          },
          "ended_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Execution end time.",
            "title": "Ended At"
          },
          "error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Error message.",
            "title": "Error"
          },
          "execution_target": {
            "$ref": "#/components/schemas/ExecutionTarget"
          },
          "executor_handle": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Executor handle.",
            "title": "Executor Handle"
          },
          "experiment_id": {
            "description": "Id of the experiment.",
            "format": "uuid",
            "title": "Experiment Id",
            "type": "string"
          },
          "id": {
            "description": "Experiment run id.",
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "number": {
            "description": "Per-experiment run number.",
            "title": "Number",
            "type": "integer"
          },
          "owner_id": {
            "description": "Id of the owning account.",
            "format": "uuid",
            "title": "Owner Id",
            "type": "string"
          },
          "progress": {
            "$ref": "#/components/schemas/ExperimentRunProgress",
            "description": "Job counts by status."
          },
          "score_baselines": {
            "description": "Whether the runner also scores originals missing scores.",
            "title": "Score Baselines",
            "type": "boolean"
          },
          "started_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Execution start time.",
            "title": "Started At"
          },
          "status": {
            "$ref": "#/components/schemas/ExperimentRunStatus",
            "description": "Run status."
          },
          "summary": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "description": "Aggregate diff, written at completion.",
            "title": "Summary"
          },
          "updated": {
            "description": "Last modification time.",
            "format": "date-time",
            "title": "Updated",
            "type": "string"
          }
        },
        "required": [
          "id",
          "owner_id",
          "experiment_id",
          "number",
          "status",
          "agent_version_id",
          "score_baselines",
          "execution_target",
          "executor_handle",
          "started_at",
          "ended_at",
          "summary",
          "error",
          "progress",
          "created",
          "updated"
        ],
        "title": "ExperimentRunResponse",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "Page_ExperimentRunResponse_": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/ExperimentRunResponse"
            },
            "title": "Items",
            "type": "array"
          },
          "page": {
            "title": "Page",
            "type": "integer"
          },
          "page_size": {
            "title": "Page Size",
            "type": "integer"
          },
          "total": {
            "title": "Total",
            "type": "integer"
          }
        },
        "required": [
          "items",
          "total",
          "page",
          "page_size"
        ],
        "title": "Page[ExperimentRunResponse]",
        "type": "object"
      },
...
      "Page_SessionResponse_": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/SessionResponse"
            },
            "title": "Items",
            "type": "array"
          },
          "page": {
            "title": "Page",
            "type": "integer"
          },
          "page_size": {
            "title": "Page Size",
            "type": "integer"
          },
          "total": {
            "title": "Total",
            "type": "integer"
          }
        },
        "required": [
          "items",
          "total",
          "page",
          "page_size"
        ],
        "title": "Page[SessionResponse]",
        "type": "object"
      },
...
      "SessionOrigin": {
        "description": "Session origin.",
        "enum": [
          "imported",
          "recorded",
          "replay"
        ],
        "title": "SessionOrigin",
        "type": "string"
      },
      "SessionProvider": {
        "description": "Session provider.",
        "enum": [
          "langfuse",
          "braintrust",
          "otlp"
        ],
        "title": "SessionProvider",
        "type": "string"
      },
...
      "SessionRunCreateRequest": {
        "additionalProperties": false,
        "description": "Session run create request.",
        "properties": {
          "agent_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Id of the agent to execute.",
            "title": "Agent Id"
          },
          "agent_version_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Id of the agent version to execute, the latest runnable version of the agent when omitted.",
            "title": "Agent Version Id"
          },
          "execution_target": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ExecutionTarget"
              },
              {
                "type": "null"
              }
            ],
            "description": "Execution target, the agent version's default when omitted."
          },
          "inputs": {
            "description": "Session inputs.",
            "title": "Inputs"
          },
          "name": {
            "anyOf": [
              {
                "maxLength": 255,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Session run name.",
            "title": "Name"
          }
        },
        "title": "SessionRunCreateRequest",
        "type": "object"
      },
...
      "SessionStatus": {
        "description": "Session status.",
        "enum": [
          "in_progress",
          "completed",
          "failed"
        ],
        "title": "SessionStatus",
        "type": "string"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "Kitaru",
    "version": "0.21.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/v1/cohorts/{cohort_id}/sessions": {
      "get": {
        "description": "List the member sessions of a cohort ordered by position.\n\nClients observe HTTP 200 on success, 404 when no cohort has this id,\nand 422 on invalid pagination parameters.\n\nArgs:\n    cohort_id: Id of the cohort.\n    service: Cohort service.\n    actor: Caller context.\n    page: Page number.\n    page_size: Page size.\n\nReturns:\n    Page of member sessions.",
        "operationId": "list_cohort_sessions_v1_cohorts__cohort_id__sessions_get",
        "parameters": [
          {
            "in": "path",
            "name": "cohort_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Cohort Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "page",
            "required": false,
            "schema": {
              "default": 1,
              "minimum": 1,
              "title": "Page",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "page_size",
            "required": false,
            "schema": {
              "default": 20,
              "maximum": 1000,
              "minimum": 1,
              "title": "Page Size",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Page_SessionResponse_"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Cohort Sessions",
        "tags": [
          "cohorts"
        ]
      }
    },
...
    "/v1/experiments/{experiment_id}/runs": {
      "get": {
        "description": "List the runs of an experiment.\n\nClients observe HTTP 200 on success, 404 when no experiment has this\nid, and 422 on invalid pagination parameters.\n\nArgs:\n    experiment_id: Id of the experiment.\n    service: Experiment run service.\n    actor: Caller context.\n    page: Page number.\n    page_size: Page size.\n\nReturns:\n    Page of experiment runs.",
        "operationId": "list_experiment_runs_v1_experiments__experiment_id__runs_get",
        "parameters": [
          {
            "in": "path",
            "name": "experiment_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Experiment Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "page",
            "required": false,
            "schema": {
              "default": 1,
              "minimum": 1,
              "title": "Page",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "page_size",
            "required": false,
            "schema": {
              "default": 20,
              "maximum": 1000,
              "minimum": 1,
              "title": "Page Size",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Page_ExperimentRunResponse_"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Experiment Runs",
        "tags": [
          "experiments"
        ]
      },
      "post": {
        "description": "Start an experiment run.\n\nCreates the run plus one pending job per cohort session.\n\nClients observe HTTP 201 on success, 404 when no experiment or agent\nversion has the referenced id, 409 when no runnable agent version\nresolves or an on demand run resolves to a version without an image,\nand 422 when the version belongs to another agent.\n\nArgs:\n    experiment_id: Id of the experiment.\n    body: Experiment run create request.\n    service: Experiment service.\n    actor: Caller context.\n\nReturns:\n    Created experiment run.",
        "operationId": "create_experiment_run_v1_experiments__experiment_id__runs_post",
        "parameters": [
          {
            "in": "path",
            "name": "experiment_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Experiment Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ExperimentRunCreateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExperimentRunResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Experiment Run",
        "tags": [
          "experiments"
        ]
      }
    },
...
    "/v1/session-runs": {
      "post": {
        "description": "Create a session run of one agent version.\n\nClients observe HTTP 201 on success, 404 when no agent version has\nthe referenced id, 409 when no runnable agent version resolves or an\non demand run resolves to a version without an image, and 422 when\nthe version belongs to another agent or the input is invalid.\n\nArgs:\n    body: Session run create request.\n    service: Job service.\n    actor: Caller context.\n\nReturns:\n    Created job.",
        "operationId": "create_session_run_v1_session_runs_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SessionRunCreateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Session Run",
        "tags": [
          "session-runs"
        ]
      }
    },
    "/v1/sessions": {
      "get": {
        "description": "List sessions.\n\nClients observe HTTP 200 on success, 404 when no agent has the filtered\nagent id, and 422 on invalid filter or pagination parameters.\n\nArgs:\n    service: Session service.\n    actor: Caller context.\n    agent_id: Filter on agent id.\n    agent_version_id: Filter on agent version id.\n    origin: Filter on session origin.\n    status: Filter on session status.\n    provider: Filter on session provider.\n    external_id: Filter on external id.\n    name: Filter on session name.\n    tag: Filter on attached tag name.\n    started_after: Earliest start time.\n    started_before: Latest start time.\n    ended_after: Earliest end time.\n    ended_before: Latest end time.\n    has_score: Filter on the presence of scores.\n    min_cost: Lowest cost.\n    max_cost: Highest cost.\n    min_total_tokens: Lowest total token count.\n    max_total_tokens: Highest total token count.\n    page: Page number.\n    page_size: Page size.\n\nReturns:\n    Page of sessions.",
        "operationId": "list_sessions_v1_sessions_get",
        "parameters": [
          {
            "in": "query",
            "name": "agent_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "uuid",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Agent Id"
            }
          },
          {
            "in": "query",
            "name": "agent_version_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "uuid",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Agent Version Id"
            }
          },
          {
            "in": "query",
            "name": "origin",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/SessionOrigin"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Origin"
            }
          },
          {
            "in": "query",
            "name": "status",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/SessionStatus"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          },
          {
            "in": "query",
            "name": "provider",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/SessionProvider"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Provider"
            }
          },
          {
            "in": "query",
            "name": "external_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "External Id"
            }
          },
          {
            "in": "query",
            "name": "name",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Name"
            }
          },
          {
            "in": "query",
            "name": "tag",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Tag"
            }
          },
          {
            "in": "query",
            "name": "started_after",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "date-time",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Started After"
            }
          },
          {
            "in": "query",
            "name": "started_before",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "date-time",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Started Before"
            }
          },
          {
            "in": "query",
            "name": "ended_after",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "date-time",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Ended After"
            }
          },
          {
            "in": "query",
            "name": "ended_before",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "date-time",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Ended Before"
            }
          },
          {
            "in": "query",
            "name": "has_score",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "boolean"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Has Score"
            }
          },
          {
            "in": "query",
            "name": "min_cost",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "number"
                },
                {
                  "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Min Cost"
            }
          },
          {
            "in": "query",
            "name": "max_cost",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "number"
                },
                {
                  "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Max Cost"
            }
          },
          {
            "in": "query",
            "name": "min_total_tokens",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Min Total Tokens"
            }
          },
          {
            "in": "query",
            "name": "max_total_tokens",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Max Total Tokens"
            }
          },
          {
            "in": "query",
            "name": "page",
            "required": false,
            "schema": {
              "default": 1,
              "minimum": 1,
              "title": "Page",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "page_size",
            "required": false,
            "schema": {
              "default": 20,
              "maximum": 1000,
              "minimum": 1,
              "title": "Page Size",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Page_SessionResponse_"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Sessions",
        "tags": [
          "sessions"
        ]
      },
      "post": {
        "description": "Create a session.\n\nClients observe HTTP 201 on success, 404 when no agent or agent version\nhas the referenced id, 409 when the provider and external id pair is\nalready registered, and 422 on invalid input.\n\nArgs:\n    body: Session create request.\n    service: Session service.\n    actor: Caller context.\n\nReturns:\n    Created session.",
        "operationId": "create_session_v1_sessions_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SessionCreateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SessionResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Session",
        "tags": [
          "sessions"
        ]
      }
    },
    "/v1/sessions/{session_id}": {
      "delete": {
        "description": "Delete a session, including its nodes and tag links.\n\nClients observe HTTP 204 on success, 404 when no session has this id,\nand 409 while the session is a member of a cohort or referenced by a\njob.\n\nArgs:\n    session_id: Id of the session.\n    service: Session service.\n    actor: Caller context.",
        "operationId": "delete_session_v1_sessions__session_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "session_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Session Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Delete Session",
        "tags": [
          "sessions"
        ]
      },
      "get": {
        "description": "Get a session by id.\n\nClients observe HTTP 200 on success and 404 when no session has this\nid.\n\nArgs:\n    session_id: Id of the session.\n    service: Session service.\n    actor: Caller context.\n\nReturns:\n    Stored session.",
        "operationId": "get_session_v1_sessions__session_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "session_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Session Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SessionResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Session",
        "tags": [
          "sessions"
        ]
      },
      "patch": {
        "description": "Update a session, finishing it when a status is set.\n\nA set status finishes the session with the request's outputs, error,\nended_at, and log_uri, and computes the rollups from its nodes. Name,\nexpected, and metadata apply to any session.\n\nClients observe HTTP 200 on success, 404 when no session has this id,\n409 when a status is set but the session is not in progress, and 422 on\ninvalid input.\n\nArgs:\n    session_id: Id of the session.\n    body: Session update request.\n    service: Session service.\n    actor: Caller context.\n\nReturns:\n    Updated session.",
        "operationId": "update_session_v1_sessions__session_id__patch",
        "parameters": [
          {
            "in": "path",
            "name": "session_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Session Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SessionUpdateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SessionResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Update Session",
        "tags": [
          "sessions"
        ]
      }
    },
...
  }
}